+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## Not released yet
 
+- 2023-08-10: Updated the adapter to align with preCICE V3. Updated the examples accordingly.
+
 ## v1.0.0
 
-- 2023-08-10: Updated the adapter to align with preCICE V3. Updated the examples accordingly.
 - 2022-09-14: The solver dummy has been cleaned up.
 - 2022-09-14: Updated CI to use containers from `precice` namespace on DockerHub. Also makes sure that `dune-precice` containers are build for tests against DuMuX's master branch.
 - 2022-09-14: Updated and fixed GitHub to build Docker containers.